Collision edits/fixes and class rearangment
Fixed some of the logic within SAP, where some boxes would get pushed down below the map or past the map in x direction. Rearranged the class Square because it was getting too confusing for me

Seems to working fine for now. Only issue ive seen now is when red and blue are falling red phases through blue and ends up below it, when it should stay above. I think this is an issue with swapping the velocities or collision check