Bug fixes
My keepInBounds fucntion was causing some errors within the SAP function. Changed it how i have it within the squaresINmotion function and things seem to be working fine.

Error for red square phasing through blue at inital start was becasue red's velocity and because of the  delata time. Time was too fast wasnt allowing for proper detection between the two. For now changed the speed of deltaTime, but I need to find a better fix