Function Fix
Fixed keepInBounds function, now checks if we are doing a horizontal or vertical check. There was an issue with vertical check where I was checking the bottom axis wrong.

IDK if i should add a bool isVertical as well, so then we can check both axis at the same time or just call the function again but with isHorizontal as false