Debuging
Trying to figure out why the red square is tunelling through the blue one. It has to do with the deltaTime because when i change the value of dt, making it slower to works fine. Need to figure out how how to prevent this

Adding print statments to see (whats going on) marvin gaye