Potential fix for movement of squrares
Made the condition to reduce the x velocity less strict. Now we only check if current square.y position is less than or equal to its groundY value and if its yMotion is false isntead of checking for vel.vy == 0. I also changed the epsilon check to 0.016f because that was the smallest value I saw when debuggin.